compression: move low level compression code to separate package
Move the compression code to a separate package. This package will
expose compression levels (useful for running experiments), while the
`sstable/block` options will remain higher level.
