+++ internal/compression/zstd_nocgo.go
@@ -1,10 +1,10 @@
-// Copyright 2021 The LevelDB-Go and Pebble Authors. All rights reserved. Use
+// Copyright 2025 The LevelDB-Go and Pebble Authors. All rights reserved. Use
 // of this source code is governed by a BSD-style license that can be found in
 // the LICENSE file.
 
 //go:build !cgo
 
-package block
+package compression
 
 import (
 	"encoding/binary"
@@ -34,7 +34,7 @@ const UseStandardZstdLib = false
 // is sufficient. The subslice `compressedBuf[:varIntLen]` should already encode
 // the length of `b` before calling Compress. It returns the encoded byte
 // slice, including the `compressedBuf[:varIntLen]` prefix.
-func (zstdCompressor) Compress(compressedBuf, b []byte) (CompressionIndicator, []byte) {
+func (zstdCompressor) Compress(compressedBuf, b []byte) []byte {
 	if len(compressedBuf) < binary.MaxVarintLen64 {
 		compressedBuf = append(compressedBuf, make([]byte, binary.MaxVarintLen64-len(compressedBuf))...)
 	}
@@ -44,7 +44,7 @@ func (zstdCompressor) Compress(compressedBuf, b []byte) (CompressionIndicator, [
 	if err := encoder.Close(); err != nil {
 		panic(err)
 	}
-	return ZstdCompressionIndicator, result
+	return result
 }
 
 func (zstdCompressor) Close() {}
@@ -57,8 +57,6 @@ type zstdDecompressor struct{}
 
 var _ Decompressor = zstdDecompressor{}
 
-// Decompress decompresses src with the Zstandard algorithm. The destination
-// buffer must already be sufficiently sized, otherwise DecompressInto may error.
 func (zstdDecompressor) DecompressInto(dst, src []byte) error {
 	// The payload is prefixed with a varint encoding the length of
 	// the decompressed block.
@@ -75,6 +73,16 @@ func (zstdDecompressor) DecompressInto(dst, src []byte) error {
 			errors.Safe(result), errors.Safe(dst))
 	}
 	return nil
+}
+
+func (zstdDecompressor) DecompressedLen(b []byte) (decompressedLen int, err error) {
+	// This will also be used by zlib, bzip2 and lz4 to retrieve the decodedLen
+	// if we implement these algorithms in the future.
+	decodedLenU64, varIntLen := binary.Uvarint(b)
+	if varIntLen <= 0 {
+		return 0, base.CorruptionErrorf("pebble: compression block has invalid length")
+	}
+	return int(decodedLenU64), nil
 }
 
 func (zstdDecompressor) Close() {}
